type error corrected

+++ alignment_algorithm/subproblem_alignment.h
@@ -8,7 +8,7 @@
 typedef struct alignment_point {
     size_t i;
     size_t j;
-    size_t t;
+    int t;
     struct alignment_point* next = NULL;
 } align;
